refactor code for readability of feature tiles
