+++ antd-mobile/src/components/featuretiles/Donation.jsx
@@ -1,0 +1,5 @@
+export default function Donation() {
+  return (
+    <div>Donation</div>
+  )
+}
