+++ antd-mobile/src/components/featuretiles/DynamicQrGenerater.jsx
@@ -1,0 +1,45 @@
+import { Button, Card } from "antd-mobile";
+import { PayCircleOutline } from "antd-mobile-icons";
+import { useNavigate } from "react-router-dom";
+
+export default function DynamicQrGenerater() {
+  const navigate = useNavigate();
+
+  const goToDonation = () => {
+    console.log("DonationPage")
+    navigate('/qr-creation');
+  };
+
+  return (
+    <Card style={{
+      marginBottom: '16px',
+      borderRadius: '8px',
+      background: 'linear-gradient(45deg, #8b0000, #b00000)',
+      color: 'white'
+    }}>
+      <div style={{
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'space-between'
+      }}>
+        <div>
+          <h3 style={{ margin: '0 0 5px 0', color: 'white' }}>Create QR Code</h3>
+          <p style={{ margin: 0, fontSize: '14px', color: 'rgba(255,255,255,0.8)' }}>
+            Your contribution helps us grow
+          </p>
+        </div>
+        <Button
+          onClick={goToDonation}
+          style={{
+            backgroundColor: 'white',
+            color: '#b00000',
+            borderColor: 'white',
+            borderRadius: '20px'
+          }}
+        >
+          <PayCircleOutline fontSize={16} /> Create
+        </Button>
+      </div>
+    </Card>
+  )
+}
